requirements.txt
dash==2.18.2
dash-bootstrap-components==1.4.1
pandas==2.2.3
plotly==5.15.0
openai==1.55.3
Flask==2.3.2
fetchai==0.1.24
python-dotenv==1.0.0
flask_cors==5.0.0
